packages/deeprag-core/pyproject.toml
[project]
name = "deeprag-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
 "diskcache>=5.6.3",
 "einops>=0.8.1",
 "filetype>=1.2.0",
 "flagembedding>=1.3.4",
 "gritlm>=1.0.2",
 "langchain>=0.3.24",
 "langchain-chroma>=0.1.2",
 "langchain-openai>=0.3.14",
 "loguru>=0.7.3",
 "nest-asyncio>=1.6.0",
 "networkx>=3.4.2",
 "numpy<2",
 "openai>=1.75.0",
 "pillow>=11.2.1",
 "pydantic>=2.11.3",
 "pypdf[crypto]>=5.4.0",
 "python-igraph>=0.11.8",
 "scipy>=1.15.2",
 "sentence-transformers>=4.1.0",
 "sqlmodel>=0.0.24",
 "tenacity>=9.1.2",
 "tiktoken>=0.9.0",
 "toml>=0.10.2",
 "tomli>=2.2.1",
 "torch>=2.6.0",
 "tqdm>=4.67.1",
 "transformers>=4.51.3",
 "trio-asyncio>=0.15.0",
 "vllm>=0.8.4",
]

[tool.uv]
[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true



